split mdf_reader.utils.auxiliary
